extract sqrt, enclose, and verb into their own files
